Separación de entidades para manejo más coherente de la información referente a combustibles
